requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
faiss-cpu
numpy
langchain-openai
python-dotenv
openai
